+++ src/app/components/price-row/price-row.module.ts
@@ -1,0 +1,12 @@
+import { NgModule } from '@angular/core';
+import {IonicModule} from '@ionic/angular';
+import {PriceRowComponent} from './price-row.component';
+
+@NgModule({
+  imports: [
+    IonicModule
+  ],
+  declarations: [PriceRowComponent],
+  exports: [PriceRowComponent]
+})
+export class PriceRowModule {}