+++ src/app/pages/reports/reports.module.ts
@@ -7,10 +7,10 @@ import { IonicModule } from '@ionic/angular';
 import { ReportsPageRoutingModule } from './reports-routing.module';
 
 import { ReportsPage } from './reports.page';
-import {InfoCardModule} from "../../components/info-card/info-card.module";
-import {EnergyBarChartModule} from "../../components/energy-bar-chart/energy-bar-chart.module";
-import {CostChartModule} from "../../components/cost-chart/cost-chart.module";
-import {GasChartModule} from "../../components/gas-chart/gas-chart.module";
+import {InfoCardModule} from '../../components/info-card/info-card.module';
+import {EnergyBarChartModule} from '../../components/energy-bar-chart/energy-bar-chart.module';
+import {CostChartModule} from '../../components/cost-chart/cost-chart.module';
+import {GasChartModule} from '../../components/gas-chart/gas-chart.module';
 
 @NgModule({
     imports: [