+++ src/app/pages/settings/prices/prices.page.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import {SettingsService} from '../../../services/settings.service';
 
 @Component({
   selector: 'app-prices',
@@ -6,10 +7,24 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./prices.page.scss'],
 })
 export class PricesPage implements OnInit {
+  public eUsagePrice: number;
+  public eProductionPrice: number;
+  public gasUsagePrice: number;
 
-  constructor() { }
+  public constructor(private readonly settings: SettingsService) { }
 
-  ngOnInit() {
+  public ngOnInit() {
+    const prices = this.settings.getPrices();
+
+    this.gasUsagePrice = prices.gas;
+    this.eProductionPrice = prices.powerProduction;
+    this.eUsagePrice = prices.powerUsage;
   }
 
+  public onUpdate(value: string, update: number) {
+    const prices = this.settings.getPrices();
+
+    prices[value] = update;
+    this.settings.setPrices(prices);
+  }
 }